% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin.R
\docType{data}
\name{builtin_registry}
\alias{builtin_registry}
\title{Built-in handler registry}
\format{
A named list of lists. The shape is
\code{builtin_registry[[mode]][[fn]] <- function(args) { ... }}.
}
\usage{
builtin_registry
}
\description{
A minimal allow-listed registry of built-in compute handlers. This exists
so the service can start and be integration-tested before package/snippet
registries are wired up.
}
\examples{
builtin_registry$builtin$score_move(list(state = list(x = 5L)))

}
\keyword{datasets}
