% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.R
\name{fallback}
\alias{fallback}
\alias{\%||\%}
\alias{`\%||\%`}
\title{Fallback operator}
\usage{
a \%||\% b
}
\arguments{
\item{a, b}{Any R objects.}
}
\value{
One of \code{a} or \code{b}.
}
\description{
Returns \code{b} when \code{a} is \code{NULL}, otherwise returns \code{a}.
}
\keyword{internal}
